refactor(client)!: introduce token caching and improve ergonomics
This commit refactors the client and authentication flow to improve ergonomics and performance.

A token caching mechanism has been introduced via `FtApiToken::try_get`. This function attempts to retrieve a valid token from a temporary file, falling back to fetching a new one if the cached token is invalid or non-existent. This reduces the number of authentication requests.

The `journals.rs` example binary has been updated to use a new `scroller` utility, simplifying pagination logic.

Additionally, extensive documentation has been added across the `auth`, `client`, and `connector` modules to improve developer experience.

BREAKING CHANGE: The `page` field in `FtApiCampusIdJournalsRequest` has been changed from `Option<u16>` to `Option<usize>` to support a larger number of pages.

+++ libft-api/src/api/campus/campus_id_journals.rs
@@ -14,7 +14,7 @@ pub struct FtApiCampusIdJournalsRequest {
     pub sort: Option<Vec<FtSortOption>>,
     pub range: Option<Vec<FtRangeOption>>,
     pub filter: Option<Vec<FtFilterOption>>,
-    pub page: Option<u16>,
+    pub page: Option<usize>,
     pub per_page: Option<u8>,
 }
 
@@ -28,6 +28,27 @@ impl<FCHC> FtClientSession<'_, FCHC>
 where
     FCHC: FtClientHttpConnector + Send + Sync,
 {
+    /// Get journals for a specific campus.
+    ///
+    /// This action requires the 'Advanced staff' role.
+    /// This resource is paginated, with a default of 30 items per page.
+    /// You have to provide parameters with FtApiCampusIdJournalsRequest structure
+    ///
+    /// # Parameters
+    ///
+    /// *   `begin_at`: **Required** (`String`). Must be before or equal to `end_at`. The date range must be 124 days maximum.
+    /// *   `end_at`: **Required** (`String`). Must be after or equal to `begin_at`. The date range must be 124 days maximum.
+    /// *   `campus_id`: **Required** (`String`). The campus ID or slug.
+    /// *   `sort`: Optional. The sort field. Sorted by `id` desc by default.
+    ///     Must be one of: `id`, `user_id`, `item_type`, `item_id`, `cursus_id`, `campus_id`, `reason`, `created_at`, `updated_at`, `event_at`, `alumni`, `closed`.
+    /// *   `filter`: Optional. Filtering on one or more fields.
+    ///     Must be one of: `id`, `user_id`, `item_type`, `item_id`, `cursus_id`, `campus_id`, `reason`, `created_at`, `updated_at`, `event_at`, `alumni`, `closed`, `event`.
+    /// *   `page[size]`: Optional (`Integer`). The number of items per page. Defaults to 30, maximum 100.
+    /// *   `page[number]`: Optional (`Integer`). The current page number.
+    ///
+    /// # Errors
+    ///
+    /// * This function will return an error if the authenticated user does not have the [`Advanced staff`] role.
     pub async fn campus_id_journals(
         &self,
         req: FtApiCampusIdJournalsRequest,
